Update README.md
PROJECT DESCRIPTION
One of the most important factors which affect our country’s economy and financial condition is the credit system governed by the banks. The process of bank credit
risk evaluation is recognized at banks across the globe. “As we know credit risk evaluation is very crucial, there is a variety of techniques are used for risk level 
calculation. In addition, credit risk is one of the main functions of the banking community.

The prediction of credit defaulters is one of the difficult tasks for any bank. But by forecasting the loan defaulters, the banks definitely may reduce their loss 
by reducing their non-profit assets, so that recovery of approved loans can take place without any loss and it can play as the contributing parameter of the 
bank statement. This makes the study of this loan approval prediction important. Machine Learning techniques are very crucial and useful in the prediction of these 
types of data.

We will be using classification algorithms such as Decision tree, Random forest, KNN, and xgboost. We will train and test the data with these algorithms. 
From this best model is selected and saved in pkl format. We will be doing flask integration and IBM deployment.